requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pypdf==3.12.0
requests==2.31.0
openai==1.11.0
numpy==1.26.4
pinecone-client==4.1.2
pytest==7.4.2
